+++ extensions.sh
@@ -1,0 +1,4 @@
+apt install php-mysqli php-curl php-zip
+service apache2 restart
+echo ""
+echo "PHP Extions were installed, Apache2 was restarted to enable the Extensions."